Refactor ClientInvoicesPanel into hook and subcomponents
Split the 2802-line client-invoices-panel into useClientInvoicesPanel
and seven focused UI components. The panel file now composes subcomponents
and re-exports NO_ENROLLMENT_OPTION_VALUE. Behavior and tests unchanged.
